fix FX29: VX*20 to compensate for size of each sprite
I'm not sure if this is correct. REG.I is supposed to point to sprite
location referenced by VX. But sprite / font data is 512bytes large,
and VX is 0-255. So clearly VX needs to be scaled by 2 or greater in order
to allow REG.I to point to full range of sprites/fonts. I scale by 20 since
wikipedia says each font is 4x5 (which also doesn't make much sense as
the drawing function draws at 8 pixels wide.