implement DXYN -- draw sprite at X,Y, N height
this initial attempt is untested. It also draws directly to VRAM
tiles. This should be changed later to modify in-memory copies of the
tiles in VRAM. CHIP is 64x32 resolution, SCHIP is 128x64, so for
compatibility reasons I'v decided that tiles onscreen will be in
sequential order. Each row of tiles onscreen will be represented by a
row of tiles in VRAM. 16 tiles per row (for a resolution of 128), 8
rows (for a resolution of 64). Just initialize the middle screen's
tiles to be sequential order $00, $01, ... $0F, $10, $11... $F0.
Then all other tiles will be tile $F1, which will just be blank.
I hypothesize that VBLANK will be used to copy these 128 tiles from
memory to VRAM