fix FX33: was always pointing to 0,0,0 bcd
I accidentally assumed a macro overwrote BC, so I pushed and popped
BC. But along the way I loaded the actual bcd table offset address
in DE instead of BC. So when I was using BC I was always using the
first address of the bcd table: aka 0,0,0. DE held the actual address
for the correct bcd values.