double horizontal screen size
turned out to be surprisingly easy. Drawing routine was modified to
treat each tile as 8x4 pixels (8 tall, 4 wide). One problem that turned
out to be harmless was the expanded size of the buffer screen. The size
expanded from 256 to 512 bytes, which pushed the full CHIP8 ram
requirements to > 4KB. Luckily the gameboy has 8KB of internal ram.