copy A-Za-z characters to vram
plus a few others after. Allows me to easily display a sentence
on-screen by simply writing the string to the BG tile-map. The character
fonts will be located at the appropriate tile within VRAM. The characters
that are copied over start with "@" but it's in the place of the blank
tile, hence it gets erased immediately