organize chip8 roms
including how they are included in the code. The ROM_COPY macro includes
code to copy the rom to ram and start the emulation.
Apparently SECTION "", ROM0 for specifying on which bank data resides
does not expand beyond 16KB unless "tinyrom" (the -t option) is passed
to rgblink to take up 32KB max. So I put all the roms under ROM0 and the
rest of the code under ROMX