switch between games by pressing SELECT
I do this by having a manual list of call <xyz> for each xyz.ch8
ROM that I include. If you want to include your own, add a
ROM_COPY <your_rom>
next to the others, and also a
call <your_rom>
with the rest of call   rom_....
And finally of course, include your file in the same folder
as all the other .ch8 roms

Known bug:
* Background on message seems to load some detail as the
game loads, and in some games, as the game plays (keypad test)
* you can only switch a game if the game itself is checking
for keypresses. Once a game ends, it typically stops looking
for keypresses, and so you won't be able to switch games then.